+++ integration_tests/suite/test_yealink.py
@@ -21,6 +21,8 @@ from .helpers.wait_strategy import PhonedEverythingUpWaitStrategy
 VENDOR = 'yealink'
 
 
+# NOTE(afournier): these tests should be removed at the same time that the deprecated directories
+# route is removed.
 class TestYealink(BasePhonedIntegrationTest):
 
     asset = 'default_config'
@@ -256,6 +258,218 @@ class TestYealink(BasePhonedIntegrationTest):
 
     def test_that_lookup_return_error_when_no_term(self):
         response = self.get_lookup_result(
+            vendor=VENDOR, xivo_user_uuid=USER_1_UUID, profile=DEFAULT_PROFILE,
+        )
+        assert_that(response.status_code, equal_to(400))
+
+
+class TestYealinkDirectories(BasePhonedIntegrationTest):
+
+    asset = 'default_config'
+    wait_strategy = PhonedEverythingUpWaitStrategy()
+
+    # Lookup
+
+    def test_that_lookup_return_no_error_when_query_ssl(self):
+        response = self.get_ssl_directories_lookup_result(
+            vendor=VENDOR,
+            xivo_user_uuid=USER_1_UUID,
+            profile=DEFAULT_PROFILE,
+            term=VALID_TERM,
+        )
+        assert_that(response.status_code, equal_to(200))
+        assert_that(
+            response.text,
+            equal_to(
+                dedent(
+                    """\
+        <?xml version="1.0" encoding="UTF-8" ?>
+        <YealinkIPPhoneDirectory>
+        <DirectoryEntry>
+          <Name>Test User1</Name>
+          <Telephone>0033123456789</Telephone>
+         </DirectoryEntry>
+        <DirectoryEntry>
+          <Name>Test User1 (mobile)</Name>
+          <Telephone>5555555555</Telephone>
+         </DirectoryEntry>
+        <DirectoryEntry>
+          <Name>Test User2</Name>
+          <Telephone>1000</Telephone>
+         </DirectoryEntry>
+        </YealinkIPPhoneDirectory>""".format(
+                        port=self.service_port(9499, 'phoned'),
+                        profile=DEFAULT_PROFILE,
+                        user_uuid=USER_1_UUID,
+                    )
+                )
+            ),
+        )
+
+    def test_that_lookup_return_no_entries_when_no_results(self):
+        response = self.get_ssl_directories_lookup_result(
+            vendor=VENDOR,
+            xivo_user_uuid=USER_1_UUID,
+            profile=DEFAULT_PROFILE,
+            term='no-result',
+        )
+        assert_that(response.status_code, equal_to(200))
+        assert_that(
+            response.text,
+            equal_to(
+                dedent(
+                    """\
+        <?xml version="1.0" encoding="UTF-8" ?>
+        <YealinkIPPhoneDirectory>
+        <DirectoryEntry>
+          <Name>No entries</Name>
+          <Telephone></Telephone>
+         </DirectoryEntry>
+        </YealinkIPPhoneDirectory>"""
+                )
+            ),
+        )
+
+    def test_that_lookup_with_limit_one_shows_first_result(self):
+        response = self.get_ssl_directories_lookup_result(
+            vendor=VENDOR,
+            xivo_user_uuid=USER_1_UUID,
+            profile=DEFAULT_PROFILE,
+            term=VALID_TERM,
+            limit=1,
+        )
+        assert_that(response.status_code, equal_to(200))
+        assert_that(
+            response.text,
+            equal_to(
+                dedent(
+                    """\
+        <?xml version="1.0" encoding="UTF-8" ?>
+        <YealinkIPPhoneDirectory>
+        <DirectoryEntry>
+          <Name>Test User1</Name>
+          <Telephone>0033123456789</Telephone>
+         </DirectoryEntry>
+        </YealinkIPPhoneDirectory>""".format(
+                        port=self.service_port(9499, 'phoned'),
+                        profile=DEFAULT_PROFILE,
+                        user_uuid=USER_1_UUID,
+                        term=VALID_TERM,
+                    )
+                )
+            ),
+        )
+
+    def test_that_lookup_with_limit_one_offset_one_shows_middle_result(self):
+        response = self.get_ssl_directories_lookup_result(
+            vendor=VENDOR,
+            xivo_user_uuid=USER_1_UUID,
+            profile=DEFAULT_PROFILE,
+            term=VALID_TERM,
+            limit=1,
+            offset=1,
+        )
+        assert_that(response.status_code, equal_to(200))
+        assert_that(
+            response.text,
+            equal_to(
+                dedent(
+                    """\
+        <?xml version="1.0" encoding="UTF-8" ?>
+        <YealinkIPPhoneDirectory>
+        <DirectoryEntry>
+          <Name>Test User1 (mobile)</Name>
+          <Telephone>5555555555</Telephone>
+         </DirectoryEntry>
+        </YealinkIPPhoneDirectory>""".format(
+                        port=self.service_port(9499, 'phoned'),
+                        profile=DEFAULT_PROFILE,
+                        user_uuid=USER_1_UUID,
+                        term=VALID_TERM,
+                    )
+                )
+            ),
+        )
+
+    def test_that_lookup_with_limit_one_offset_two_shows_last_result(self):
+        response = self.get_ssl_directories_lookup_result(
+            vendor=VENDOR,
+            xivo_user_uuid=USER_1_UUID,
+            profile=DEFAULT_PROFILE,
+            term=VALID_TERM,
+            limit=1,
+            offset=2,
+        )
+        assert_that(response.status_code, equal_to(200))
+        assert_that(
+            response.text,
+            equal_to(
+                dedent(
+                    """\
+        <?xml version="1.0" encoding="UTF-8" ?>
+        <YealinkIPPhoneDirectory>
+        <DirectoryEntry>
+          <Name>Test User2</Name>
+          <Telephone>1000</Telephone>
+         </DirectoryEntry>
+        </YealinkIPPhoneDirectory>""".format(
+                        port=self.service_port(9499, 'phoned'),
+                        profile=DEFAULT_PROFILE,
+                        user_uuid=USER_1_UUID,
+                        term=VALID_TERM,
+                    )
+                )
+            ),
+        )
+
+    def test_lookup_translation_fr(self):
+        response = self.get_ssl_directories_lookup_result(
+            vendor=VENDOR,
+            xivo_user_uuid=USER_1_UUID,
+            profile=DEFAULT_PROFILE,
+            term='no-result',
+            headers={'Accept-Language': 'gibberish,fr-CA,*q=nothing'},
+        )
+        assert_that(response.status_code, equal_to(200))
+        assert_that(
+            response.text,
+            equal_to(
+                dedent(
+                    """\
+        <?xml version="1.0" encoding="UTF-8" ?>
+        <YealinkIPPhoneDirectory>
+        <DirectoryEntry>
+          <Name>Aucune entrée</Name>
+          <Telephone></Telephone>
+         </DirectoryEntry>
+        </YealinkIPPhoneDirectory>"""
+                )
+            ),
+        )
+
+    def test_that_lookup_return_no_error_when_query(self):
+        response = self.get_directories_lookup_result(
+            vendor=VENDOR,
+            xivo_user_uuid=USER_1_UUID,
+            profile=DEFAULT_PROFILE,
+            term=VALID_TERM,
+        )
+        assert_that(response.status_code, equal_to(200))
+
+    def test_that_lookup_return_error_when_no_xivo_user_uuid(self):
+        response = self.get_directories_lookup_result(
+            vendor=VENDOR, profile=DEFAULT_PROFILE, term=VALID_TERM
+        )
+        assert_that(response.status_code, equal_to(400))
+
+    def test_that_lookup_return_error_when_invalid_user_uuid(self):
+        response = self.get_directories_lookup_result(
+            vendor=VENDOR, profile='a', xivo_user_uuid='invalid', term=VALID_TERM
+        )
+        assert_that(response.status_code, equal_to(404))
+
+    def test_that_lookup_return_error_when_no_term(self):
+        response = self.get_directories_lookup_result(
             vendor=VENDOR, xivo_user_uuid=USER_1_UUID, profile=DEFAULT_PROFILE,
         )
         assert_that(response.status_code, equal_to(400))