+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Changelog
   * GET `/0.1/yealink/user_service/forward_busy`
   * GET `/0.1/yealink/user_service/forward_unconditional`
   * GET `/0.1/yealink/user_service/forward_noanswer`
+  * GET `/0.1/yealink/directories/lookup/<profile>`
 
 20.03
 -----